# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FutureStarr API",
        "description": "FutureStar API Swagger Documentation. IF you want to use postman just import documetation json file",
        "contact": {
            "email": "admin@admin.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/public-profile/{public_profile_id}/profile": {
            "get": {
                "tags": [
                    "User Module"
                ],
                "summary": "Manage Public Profile",
                "description": "Manage Public Profile",
                "operationId": "ManagePublicProfile",
                "parameters": [
                    {
                        "name": "public_profile_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "stars-band-f1263317c8d"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/public-profile/buyer-futurestarr-e77a3cadfc4/profile": {
            "get": {
                "tags": [
                    "User Module"
                ],
                "summary": "Seller Public Profile",
                "description": "Seller Public Profile",
                "operationId": "SellerPublicProfile",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/v1/social-buzz-awards/{post_id}/listing": {
            "get": {
                "tags": [
                    "SocialBuzz"
                ],
                "summary": "Get Social Buzz Awards",
                "description": "Get Social Buzz Awards",
                "operationId": "GetSocialBuzzAwards",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "243"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}